+++ tests/Unit/Console/UpgradeDatabaseTest.php
@@ -3,12 +3,13 @@
 namespace Tests\Unit\Console;
 
 use DB;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class UpgradeDatabaseTest extends TestCase
 {
-    use WithFaker;
+    use DatabaseMigrations, WithFaker;
 
     /**
      * Test upgrade with v1 database